+++ src/BubbleScript.js
@@ -18,256 +18,9 @@ const Macro = require("./o/macro");
 const tokenize = require("./f/tokenize");
 const parse = require("./f/parse");
 const { ėval, ëval } = require("./f/eval");
+const rootBinding = require("./o/root_binding");
 
-// Makes a Bubblescript function from a
-// Javascript function.
-// Params:
-//   q: A Javascript function that will be
-//   called for this function.
-// Returns an annonomous function that is
-// sutible for use with Bubblescript.
-// #coreUtilityFunction
-// TODO: Create tests for mkfn.
-function mkfn(q) {
-  return function (params) {
-    // Handel & expansion.
-    // let splits = params.split(sAmp);
-    // console.log("hi", splits);
-    // if (splits.count() > 1) {
-    //   params = splits.first.conj(splits.rest.head);
-    //   console.log(params);
-    // }
-
-    return q.call(this, params.mapEval(this));
-  }
-}
-
-// A man walks into a bar. Bartender says
-// what'll you have?  The man says,
-// something strong,  my head is killing
-// me. 🍸
-const rootBinding = {
-  console: console,
-  require: mkfn(o => require(...o)),
-  __dirname: __dirname,
-
-  muf: function([key,val]) {
-    return this[key.toString()]
-      = ëval(this, val);
-  },
-
-  muf: function(args) {
-    let key = args.peek();
-    let val = args.pop();
-
-    // If the key turns out to be a list, then
-    // we do a function definition using the
-    // first item of the list as the key and the
-    // rest as the paramter list, otherwise do a
-    // normal key value definition.
-    if (key instanceof List) {
-      let name = key.peek().toString();
-      return this[key.peek().toString()]
-        = new Fn(this, key.pop(), val, { name });
-    } else {
-      return this[key.toString()]
-        = ëval(this, val.peek());
-    }
-  },
-
-  // fn: function([caret, stic]) {
-  //   return new Fn(this, caret, stic);
-  // },
-  // fn: function(_) {
-  //   // console.log(_);
-  //   let binding = this;
-  //   let caret = _.peek();
-  //   let stic  = _.pop();
-  //   return new Fn(binding, caret, stic);
-  // },
-
-  fn: function(args) {
-    return new Fn(this, args.first.toList(), args.rest)
-  },
-
-  macro: function(args) {
-    return new Macro(this, args.first, args.rest)
-  },
-
-  jsfn: function(args) {
-    let binding = this, x, fn;
-    x = args.push(§ymbol.for('fn'));
-    fn = ëval(binding, x);
-    return function(...args) {
-      return fn.invoke(List.from(args));
-    }
-  },
-
-  let: function([x,...xx]) {
-    let binding = Object.create(this);
-    x = x.invert();
-    while (!x.isEmpty) {
-      let k,w;
-      k = x.peek();
-      x = x.pop();
-      w = x.peek();
-      x = x.pop();
-      binding[k] = ëval(binding, w);
-    }
-    return xx.map(z =>
-      ëval(binding, z)).pop();
-  },
-
-  if: function([c,t,f]) {
-    return ëval(this,
-      ëval(this, c) ? t : f);
-  },
-
-  unless: function([c,f,t]) {
-    return ëval(this,
-      ëval(this, c) ? t : f);
-  },
-
-  blert: function(msgs) {
-    alert(this.concat(msgs));
-  },
-
-  expandmacro: function([m,n]) {
-    return ëval(this,m).expand(this, n);
-  },
-
-  loop: function([x,...xx]) {
-    var binding = Object.create(this),
-      m, recurCalled;
-
-    x = x.invert();
-    while (!x.isEmpty) {
-      let k,v;
-      k = x.peek();
-      x = x.pop();
-      v = x.peek();
-      x = x.pop();
-      binding[k] = ëval(binding, v);
-    }
-
-    binding.recur = function([a]) {
-      a = a.invert();
-      while (!a.isEmpty) {
-        let k,w;
-        k = a.peek();
-        a = a.pop();
-        w = a.peek();
-        a = a.pop();
-        binding[k] = ëval(binding, w);
-      }
-      recurCalled = true;
-    };
-
-    do {
-      recurCalled = false;
-      m = xx.map(z =>
-        ëval(binding, z)).pop();
-    } while(recurCalled);
-    return m;
-  },
-
-  list: mkfn(function(args) {
-    return args;
-  }),
-
-  vector: mkfn(function(args) {
-    return args.toVector();
-  }),
-
-  obj: mkfn(function(list) {
-    return list.partition(2).reduce(
-      function(memo, [key, val]) {
-        memo[key] = val;
-        return memo;
-      }, {});
-  }),
-
-  // obj: mkfn(function(list) {
-  //   return list.toObject();
-  // }),
-
-  do: function(args) {
-    return args.eval(this);
-  },
-
-  eval: mkfn(function(args) {
-    return args.eval(this);
-  }),
-
-  send: mkfn(function([a,b,...c]) {
-    if (b.key)
-      b = b.key;
-    if (c.length > 0) {
-      return a[b](...c);
-    } else
-      return a[b]();
-  }),
-  // get: mkfn(function(args) {
-  //    return args.reduce(
-  //       (a,b) => a ? a[b] : b);
-  // }),
-  get: mkfn(function(yeahyeahyeahs) {
-     return yeahyeahyeahs.reduce(
-        (memo,key) => memo && memo[key]);
-  }),
-  export: mkfn(function([ca,nd,y]) {
-    return ca[nd] = y;
-  }),
-  print: mkfn(function(vals) {
-    return vals.each(function(value) {
-      document.body.append(value);
-    });
-  }),
-  "+": mkfn(function(a) {
-    return a.reduce((a,b) => a+b);
-  }),
-  "-": mkfn(function(a) {
-    return a.reduce((a,b) => a-b);
-  }),
-  "*": mkfn(function(a) {
-    return a.reduce((a,b) => a*b);
-  }),
-  "/": mkfn(function(a) {
-    return a.reduce((a,b) => a/b);
-  }),
-  "=": mkfn(function([a, b]) {
-    return a == b;
-  }),
-  not: mkfn(function([y]) {
-    return !y;
-  }),
-  and: mkfn(function(a) {
-    return a.reduce((a,b) => a && b);
-  }),
-  or: mkfn(function(_) {
-    return _.reduce((a,b) => a || b);
-  }),
-  '>': mkfn(([a,b]) => {
-    return a > b;
-  }),
-  '<': mkfn(([a,b]) => {
-    return a < b;
-  }),
-  parse: mkfn(function([s]) {
-    return parse(s);
-  }),
-  concat: mkfn(function(eeks) {
-    return eeks.join('');
-  }),
-  "new": mkfn(function([m,n]) {
-      return new m(...n.toArray());
-  }),
-};
-
-// Alias muf to 🫧
-rootBinding["🫧"] = rootBinding.muf;
-
-AbstractList.configure({ ëval: ëval });
+AbstractList.configure({ ëval });
 
 (function() {
   let bnd = rootBinding;
